feat: add reusable dialog

+++ project-management-app/src/app/project-management-app/pages/main-route/main-route.component.ts
@@ -6,6 +6,7 @@ import { Subscription } from 'rxjs';
 import * as fromApp from '../../../store/app.reducer';
 import * as DashboardActions from './store/dashboard.actions';
 import { Store } from '@ngrx/store';
+import { ConfirmDialogService } from 'src/app/shared/services/confirm-dialog.service';
 
 @Component({
   selector: 'app-main-route',
@@ -15,12 +16,21 @@ import { Store } from '@ngrx/store';
 export class MainRouteComponent implements OnInit {
   public boards!: Dashboard[];
   public boardForm!: FormGroup;
+  selectedId!: string;
+
   storeSub!: Subscription;
   isLoadingBoard = false;
   isLoadingDashboard = false;
+  isUpdateMode = '';
+
+  updateTitle!: string;
+  updateContent!: string;
+
+  errorMsg!: string | null;
 
   constructor(
     private store: Store<fromApp.AppState>,
+    private dialogService: ConfirmDialogService
     ) {}
   ngOnInit(): void {
     this.boardForm = new FormGroup({
@@ -36,6 +46,7 @@ export class MainRouteComponent implements OnInit {
       this.boards = dashState.boards;
       this.isLoadingDashboard = dashState.getLoading;
       this.isLoadingBoard = dashState.postLoading;
+      this.errorMsg = dashState.dashboardError;
     })
   }
 
@@ -47,9 +58,38 @@ export class MainRouteComponent implements OnInit {
   }
 
   deleteBoard(id: string) {
-    this.store.dispatch(new DashboardActions.DeleteBoardStart(id));
+    this.dialogService.confirmDialog({
+      title: 'Delete',
+      message: 'Are you sure to delete board?',
+      cancelText: 'Cancel',
+      confirmText: 'Confirm',
+    }).subscribe((response: boolean) => {
+      if(response) {
+        this.store.dispatch(new DashboardActions.DeleteBoardStart(id));
+      }
+    })
+  }
+
+  selectBoard(board: Dashboard) {
+    this.selectedId = board.id;
+    this.updateTitle = board.title;
+    this.updateContent = board.description;
+  }
+
+  quitSelected() {
+    this.selectedId = '';
   }
 
   updateBoard(id: string) {
+    const title = this.updateTitle;
+    const description = this.updateContent;
+    this.store.dispatch(
+      new DashboardActions.UpdateBoardStart({
+        id,
+        title,
+        description
+      })
+    )
+    this.quitSelected();
   }
 }
